Moved user editable configuration to `UserConfig.py`
